+++ src/scoring/CMakeLists.txt
@@ -3,6 +3,9 @@ add_library(osh_scoring
     runtime/osh_scoring_accumulator.c
     runtime/osh_scoring_compile.c
     runtime/osh_scoring_estimator.c
+    runtime/osh_scoring_estimator_common.c
+    runtime/osh_scoring_estimator_point.c
+    runtime/osh_scoring_estimator_step.c
     runtime/osh_scoring_filter_runtime.c
     runtime/osh_scoring_postprocess.c
     runtime/osh_scoring_shadow.c
