+++ src/scoring/runtime/osh_scoring_step.c
@@ -7,20 +7,11 @@
 #include "common/osh_vect.h"
 #include "common/raytrace/osh_raytrace.h"
 #include "common/raytrace/osh_raytrace_cyl.h"
-#include "material/runtime/osh_material_runtime.h"
 #include "scoring/runtime/osh_scoring_estimator.h"
-#include "scoring/runtime/osh_scoring_kernels.h"
 #include "scoring/runtime/osh_scoring_step_internal.h"
 
 static int axis_index(struct osh_scoring_geometry_runtime const *geo, char const *label);
 static void step_scoring_segment(struct step const *st, double dir_out[3], double *len_out);
-static int find_proj_idx(struct osh_material_runtime const *tables, unsigned int z, size_t *proj_idx_out);
-static double
-compute_step_let(struct osh_scoring_runtime const *rt, struct particle const *part, struct step const *st);
-
-/* All per-estimator score_step_* deposit handlers are declared in
- * osh_scoring_step_internal.h and registered in osh_scoring_estimator.c; they are
- * extern so the estimator registry can reference them. */
 
 /* Book one group's step deposit via its estimator handler.  An unknown or
  * non-step estimator returns OSH_ENOTSUP, matching the old switch default. */
@@ -253,45 +244,6 @@ static void step_scoring_segment(struct step const *st, double dir_out[3], doubl
     *len_out = st->ds;
 }
 
-/** @brief Relativistic speed β = v/c from total kinetic energy and rest mass [MeV/c²]. */
-static inline double particle_beta(double e_kin_mev, double rest_mass_mev) {
-    double gamma_inv = rest_mass_mev / (e_kin_mev + rest_mass_mev);
-    return sqrt(1.0 - (gamma_inv * gamma_inv));
-}
-
-/**
- * @brief Barkas effective charge: z_eff = z·(1 − exp(−125·β·z^(−2/3))).
- * @pre   z > 0
- */
-static inline double particle_zeff(int z, double beta) {
-    double cbrt_z = cbrt((double) z);
-    return z * (1.0 - exp(-125.0 * beta / (cbrt_z * cbrt_z)));
-}
-
-/**
- * @brief Dose/track quality factor (z_eff/β)² [dimensionless].
- * @pre   z > 0 (delegates to particle_zeff)
- */
-static inline double particle_qeff(int z, double beta) {
-    double zeff = particle_zeff(z, beta);
-    return (zeff * zeff) / (beta * beta);
-}
-
-/**
- * @brief Find the column index for projectile atomic number z in the SP table.
- * @return 1 if found and proj_idx_out is set; 0 if z has no entry.
- */
-static int find_proj_idx(struct osh_material_runtime const *tables, unsigned int z, size_t *proj_idx_out) {
-    size_t i;
-    for (i = 0; i < tables->nprojectiles; ++i) {
-        if (tables->projectile_z[i] == z) {
-            *proj_idx_out = i;
-            return 1;
-        }
-    }
-    return 0;
-}
-
 /** @brief Return the index of the named axis in geo, or -1 if not found. */
 static int axis_index(struct osh_scoring_geometry_runtime const *geo, char const *label) {
     size_t i;
@@ -409,897 +361,5 @@ enum osh_status cyl_geometry_to_grid(struct osh_scoring_geometry_runtime const *
     grid->n[1] = 1u;
     grid->n[2] = (size_t) geo->axes[iz].nbins;
     grid->tile_order = OSH_RAYTRACE_GRID_TILE_ORDER_DEFAULT;
-    return OSH_OK;
-}
-
-/* ---- Differential scoring helpers ---------------------------------------- */
-
-/* Compute the 0-based bin index for a differential axis value.
- * Returns diff_nbins (out-of-range sentinel) when val is outside [lo, hi). */
-static inline size_t diff_axis_bin(struct osh_scoring_page_runtime const *page, double val) {
-    double frac; /* normalised position within [lo, hi) */
-    size_t bin;  /* candidate bin index */
-
-    if (!(val > page->diff_lo) || !(val < page->diff_hi)) {
-        return page->diff_nbins; /* sentinel: out of range */
-    }
-    if (page->diff_log) {
-        frac = log10(val / page->diff_lo) / log10(page->diff_hi / page->diff_lo);
-    } else {
-        frac = (val - page->diff_lo) / (page->diff_hi - page->diff_lo);
-    }
-    bin = (size_t) floor(frac * (double) page->diff_nbins);
-    return (bin < page->diff_nbins) ? bin : page->diff_nbins - 1u;
-}
-
-/* Compute LET [MeV/cm] in an explicit medium at step midpoint.
- * Uses the SP table; returns 0 when tables are unavailable or particle is neutral.
- * ovr_medium selects which row in mat_tables to look up (may differ from st->medium).
- * ovr_rho is the density used to convert SP [MeV·cm²/g] → LET [MeV/cm]. */
-static double compute_step_let_medium(struct osh_scoring_runtime const *rt,
-                                      struct particle const *part,
-                                      struct step const *st,
-                                      size_t ovr_medium,
-                                      double ovr_rho) {
-    double mean_energy;
-    double e_per_nuc;
-    size_t proj_idx;
-
-    if (part->z == 0 || part->a == 0 || !rt->mat_tables) {
-        return 0.0;
-    }
-    mean_energy = 0.5 * (st->p[3] + st->q[3]);
-    e_per_nuc = mean_energy / (double) part->a;
-    if (!find_proj_idx(rt->mat_tables, (unsigned int) part->z, &proj_idx)) {
-        return 0.0;
-    }
-    return osh_material_runtime_sp_lookup(rt->mat_tables, ovr_medium, proj_idx, e_per_nuc) * ovr_rho;
-}
-
-/* Compute the differential LET/DEDX axis value [MeV/cm] with an optional
- * per-axis Settings override.
- *
- * Semantics:
- *   no override                -> rho_transport * S_transport(E)
- *   density override only      -> rho_override  * S_transport(E)
- *   medium override only       -> rho_medium    * S_medium(E)
- *   medium + density override  -> rho_override  * S_medium(E)
- *
- * When a different medium is requested we need SP tables for that medium. If
- * only the density is overridden and no tables are available, fall back to the
- * transport-medium LET/de-ds and scale it by rho_override / rho_transport. */
-static double compute_step_let_with_override(struct osh_scoring_runtime const *rt,
-                                             struct particle const *part,
-                                             struct step const *st,
-                                             struct osh_scoring_page_override const *ovr) {
-    double let;
-    double rho;
-
-    if (!ovr || (!ovr->has_medium && !ovr->has_density_g_cm3)) {
-        return compute_step_let(rt, part, st);
-    }
-
-    if (ovr->has_medium && ovr->medium >= 0) {
-        rho = ovr->has_density_g_cm3 ? ovr->density_g_cm3
-                                     : (rt->mat_tables ? rt->mat_tables->rho[(size_t) ovr->medium] : st->rho);
-        return compute_step_let_medium(rt, part, st, (size_t) ovr->medium, rho);
-    }
-
-    rho = ovr->density_g_cm3;
-    if (rt->mat_tables && st->medium >= 0) {
-        return compute_step_let_medium(rt, part, st, (size_t) st->medium, rho);
-    }
-
-    let = compute_step_let(rt, part, st);
-    if (!(let > 0.0) || !(st->rho > 0.0)) {
-        return 0.0;
-    }
-    return let * (rho / st->rho);
-}
-
-/* Compute LET in the transport medium [MeV/cm] at step midpoint.
- * Uses the SP table at mean step energy; falls back to de/ds when tables are unavailable.
- * No per-page Settings override — callers that need override apply it themselves. */
-static double
-compute_step_let(struct osh_scoring_runtime const *rt, struct particle const *part, struct step const *st) {
-    double mean_energy; /* kinetic energy at step midpoint [MeV] */
-    double e_per_nuc;   /* mean_energy / A [MeV/u] */
-    size_t proj_idx;    /* projectile row in SP table */
-
-    if (part->z == 0 || part->a == 0 || !(st->rho > 0.0)) {
-        return 0.0;
-    }
-    if (rt->mat_tables && st->medium >= 0) {
-        mean_energy = 0.5 * (st->p[3] + st->q[3]);
-        e_per_nuc = mean_energy / (double) part->a;
-        if (find_proj_idx(rt->mat_tables, (unsigned int) part->z, &proj_idx)) {
-            return osh_material_runtime_sp_lookup(rt->mat_tables, (size_t) st->medium, proj_idx, e_per_nuc) * st->rho;
-        }
-    }
-    return (st->ds > 0.0) ? st->de / st->ds : 0.0;
-}
-
-/* Compute (z_eff/β)² at step midpoint.
- * Returns 0 for neutrals or when tables are unavailable. */
-static double
-compute_step_qeff(struct osh_scoring_runtime const *rt, struct particle const *part, struct step const *st) {
-    double mean_energy; /* kinetic energy at step midpoint [MeV] */
-    double beta;        /* particle velocity / c */
-    size_t proj_idx;    /* projectile row in table (provides rest mass) */
-
-    if (part->z == 0) {
-        return 0.0;
-    }
-    if (!rt->mat_tables) {
-        return 0.0;
-    }
-    if (!find_proj_idx(rt->mat_tables, (unsigned int) part->z, &proj_idx)) {
-        return 0.0;
-    }
-    mean_energy = 0.5 * (st->p[3] + st->q[3]);
-    beta = particle_beta(mean_energy, rt->mat_tables->projectile_mass_mev[proj_idx]);
-    if (!(beta > 0.0)) {
-        return 0.0;
-    }
-    return particle_qeff((int) part->z, beta);
-}
-
-/* Compute the diff axis value for a page given the current step.
- * Returns 0 and sets *ok=0 when the value cannot be determined
- * (e.g. LET for a neutral particle). */
-static double diff_step_val(struct osh_scoring_page_runtime const *page,
-                            struct osh_scoring_runtime const *rt,
-                            struct particle const *part,
-                            struct step const *st,
-                            int *ok) {
-    double mean_ekin; /* kinetic energy at midpoint [MeV] */
-
-    *ok = 1;
-    mean_ekin = 0.5 * (st->p[3] + st->q[3]);
-    switch (page->diff_kind) {
-    case OSH_SCORING_DIFF_EKIN:
-        return mean_ekin;
-    case OSH_SCORING_DIFF_ENUC:
-        if (part->a <= 0) {
-            *ok = 0;
-            return 0.0;
-        }
-        return mean_ekin / (double) part->a;
-    case OSH_SCORING_DIFF_EAMU:
-        if (part->a <= 0) {
-            *ok = 0;
-            return 0.0;
-        }
-        return mean_ekin / (double) part->a; /* same as ENUC for integer A */
-    case OSH_SCORING_DIFF_LET: {
-        double let = page->has_diff_sset ? compute_step_let_with_override(rt, part, st, &page->diff_sset)
-                                         : compute_step_let(rt, part, st);
-        if (!(let > 0.0)) {
-            *ok = 0;
-            return 0.0;
-        }
-        return let;
-    }
-    case OSH_SCORING_DIFF_QEFF: {
-        double qeff = compute_step_qeff(rt, part, st);
-        if (!(qeff > 0.0)) {
-            *ok = 0;
-            return 0.0;
-        }
-        return qeff;
-    }
-    default:
-        *ok = 0;
-        return 0.0;
-    }
-}
-
-/* Same as diff_axis_bin but operates on the second differential axis (diff2_*) fields.
- * Returns diff2_nbins as an out-of-range sentinel when val is outside [diff2_lo, diff2_hi). */
-static inline size_t diff2_axis_bin(struct osh_scoring_page_runtime const *page, double val) {
-    double frac; /* normalised position within [diff2_lo, diff2_hi) */
-    size_t bin;  /* candidate bin index */
-
-    if (!(val > page->diff2_lo) || !(val < page->diff2_hi)) {
-        return page->diff2_nbins; /* out-of-range sentinel */
-    }
-    if (page->diff2_log) {
-        frac = log10(val / page->diff2_lo) / log10(page->diff2_hi / page->diff2_lo);
-    } else {
-        frac = (val - page->diff2_lo) / (page->diff2_hi - page->diff2_lo);
-    }
-    bin = (size_t) floor(frac * (double) page->diff2_nbins);
-    return (bin < page->diff2_nbins) ? bin : page->diff2_nbins - 1u;
-}
-
-/* Same as diff_step_val but dispatches on page->diff2_kind.
- * Returns 0 and sets *ok=0 when the value cannot be determined
- * (e.g. LET for a neutral particle, or unknown kind). */
-static double diff2_step_val(struct osh_scoring_page_runtime const *page,
-                             struct osh_scoring_runtime const *rt,
-                             struct particle const *part,
-                             struct step const *st,
-                             int *ok) {
-    double mean_ekin;
-
-    *ok = 1;
-    mean_ekin = 0.5 * (st->p[3] + st->q[3]);
-    switch (page->diff2_kind) {
-    case OSH_SCORING_DIFF_EKIN:
-        return mean_ekin;
-    case OSH_SCORING_DIFF_ENUC:
-    case OSH_SCORING_DIFF_EAMU:
-        if (part->a <= 0) {
-            *ok = 0;
-            return 0.0;
-        }
-        return mean_ekin / (double) part->a;
-    case OSH_SCORING_DIFF_LET: {
-        double let = page->has_diff2_sset ? compute_step_let_with_override(rt, part, st, &page->diff2_sset)
-                                          : compute_step_let(rt, part, st);
-        if (!(let > 0.0)) {
-            *ok = 0;
-            return 0.0;
-        }
-        return let;
-    }
-    case OSH_SCORING_DIFF_QEFF: {
-        double qeff = compute_step_qeff(rt, part, st);
-        if (!(qeff > 0.0)) {
-            *ok = 0;
-            return 0.0;
-        }
-        return qeff;
-    }
-    default:
-        *ok = 0;
-        return 0.0;
-    }
-}
-
-/* ---- Shared estimator-driver helpers -------------------------------------
- * The score_step_* / score_point_* handlers below are "drivers": they gather the
- * per-step / per-page scalars, walk the pages in one score group, then book a
- * tiny pure kernel (osh_scoring_kernels.h) per deposit.  These helpers hold the
- * machinery shared by extensive, single-valued score quantities such as ENERGY,
- * FLUENCE, DOSE, and DOSEGY.  Averaged quantities (DLET/TLET/DQEFF/TQEFF) keep
- * their own two-pass loops because compile-time validation rejects differential
- * axes on those pages. */
-
-/* Resolve the per-page differential-axis bin offsets for this step.
- *
- * LET and QEFF appear here as axis values (Diff1Type/Diff2Type) for spectra of
- * ENERGY/FLUENCE/DOSE/DOSEGY.  They are not score quantities in this helper.
- * Returns 1 with *db / *db2 set (each 0 when inactive), or 0 to skip the page
- * when the axis value is undefined (e.g. LET for a neutral) or out of range. */
-static int resolve_diff_bins(struct osh_scoring_page_runtime const *page,
-                             struct osh_scoring_runtime const *rt,
-                             struct particle const *part,
-                             struct step const *st,
-                             size_t *db,
-                             size_t *db2) {
-    int dv_ok;
-    double dv;
-
-    *db = 0u;
-    *db2 = 0u;
-    if (page->diff_nbins > 0u) {
-        dv = diff_step_val(page, rt, part, st, &dv_ok);
-        if (!dv_ok) {
-            return 0;
-        }
-        *db = diff_axis_bin(page, dv);
-        if (*db >= page->diff_nbins) {
-            return 0;
-        }
-    }
-    if (page->diff2_nbins > 0u) {
-        dv = diff2_step_val(page, rt, part, st, &dv_ok);
-        if (!dv_ok) {
-            return 0;
-        }
-        *db2 = diff2_axis_bin(page, dv);
-        if (*db2 >= page->diff2_nbins) {
-            return 0;
-        }
-    }
-    return 1;
-}
-
-/* Flat accumulator index of spatial bin @p spatial_idx at diff offsets (db, db2);
- * an inactive differential axis has offset 0 so its term vanishes. */
-static inline size_t flat_bin(struct osh_scoring_page_runtime const *page, size_t spatial_idx, size_t db, size_t db2) {
-    return spatial_idx + (db * page->diff_stride) + (db2 * page->diff2_stride);
-}
-
-/* Book one crossing into a two-pass average page (LET/Qeff):
- * data += value·weight (numerator), data2 += weight (denominator). */
-static inline void deposit_two_pass(struct osh_scoring_accumulator *acc, size_t idx, double value, double weight) {
-    osh_score_deposit(acc->data, idx, value * weight);
-    osh_score_deposit(acc->data2, idx, weight);
-}
-
-/* Per-step projectile stopping-power context for DOSE dose-to-medium overrides.
- *
- * The expensive/common lookup is S(transport medium, E) for this projectile and
- * step midpoint.  Each page may then request a different scoring medium through
- * Settings; dose_sp_ratio() only needs one additional lookup for that page and
- * returns S(scoring medium, E) / S(transport medium, E). */
-struct dose_sp_ctx {
-    int have_proj;    /* projectile has an SP-table entry in the transport medium */
-    size_t proj_idx;  /* projectile index into the SP tables */
-    double sp_tr;     /* S(transport medium, E) [MeV·cm²/g] */
-    double e_per_nuc; /* mean step energy per nucleon [MeV/u] */
-};
-
-static struct dose_sp_ctx
-dose_sp_gather(struct osh_scoring_runtime const *rt, struct particle const *part, struct step const *st) {
-    struct dose_sp_ctx c;
-    double mean_energy;
-    struct osh_material_runtime const *mat = rt->mat_tables;
-
-    c.have_proj = 0;
-    c.proj_idx = 0;
-    c.sp_tr = 0.0;
-    c.e_per_nuc = 0.0;
-    if (mat && part->z > 0 && part->a > 0 && st->medium >= 0) {
-        /* Use the same representative energy as the LET/QEFF axis code: the
-         * midpoint between pre-step and post-step kinetic energy. */
-        mean_energy = 0.5 * (st->p[3] + st->q[3]);
-        c.e_per_nuc = mean_energy / (double) part->a;
-        if (find_proj_idx(mat, (unsigned int) part->z, &c.proj_idx)) {
-            c.have_proj = 1;
-            /* Baseline stopping power in the actual transport medium.  If this
-             * is zero, a dose-to-medium ratio is undefined and contributes 0. */
-            c.sp_tr = osh_material_runtime_sp_lookup(mat, (size_t) st->medium, c.proj_idx, c.e_per_nuc);
-        }
-    }
-    return c;
-}
-
-/* Dose-to-medium stopping-power ratio S(ovr)/S(tr) for a page's medium override;
- * 1.0 when there is no override (density-only overrides are Fano-invariant). */
-static double dose_sp_ratio(struct dose_sp_ctx const *c,
-                            struct osh_scoring_runtime const *rt,
-                            struct osh_scoring_page_runtime const *page) {
-    double sp_ovr;
-
-    if (c->have_proj && page->has_sset && page->sset.has_medium && page->sset.medium >= 0) {
-        /* Page-local Settings override: convert dose-to-transport into
-         * dose-to-medium by multiplying each raw page deposit by this ratio. */
-        sp_ovr = osh_material_runtime_sp_lookup(rt->mat_tables, (size_t) page->sset.medium, c->proj_idx, c->e_per_nuc);
-        if (c->sp_tr > 0.0) {
-            return sp_ovr / c->sp_tr;
-        }
-        return 0.0;
-    }
-    return 1.0;
-}
-
-/**
- * @brief Accumulate energy deposition [MeV] into the ENERGY scorer pages.
- *
- * Distributes st->de proportionally to path length in each crossed voxel.
- */
-enum osh_status score_step_energy(struct osh_scoring_runtime const *rt,
-                                  struct osh_scoring_accumulator *acc_set,
-                                  struct osh_scoring_geometry_score_group const *group,
-                                  struct osh_voxel_crossing const *crossings,
-                                  size_t ncross,
-                                  struct particle const *part,
-                                  struct step const *st,
-                                  double score_len) {
-    size_t i;
-    size_t j;
-    size_t db;
-    size_t db2;
-    struct osh_scoring_page_runtime const *page;
-    struct osh_scoring_accumulator *acc;
-
-    for (i = 0; i < group->npages; ++i) {
-        page = &rt->pages[group->first_page + i];
-        acc = &acc_set[group->first_page + i];
-        if (!osh_scoring_page_passes_filters(page, part, st)) {
-            continue;
-        }
-        if (!resolve_diff_bins(page, rt, part, st, &db, &db2)) {
-            continue;
-        }
-        for (j = 0; j < ncross; ++j) {
-            if (crossings[j].idx >= page->diff_stride) {
-                return OSH_ESTATE;
-            }
-            osh_score_deposit(acc->data,
-                              flat_bin(page, crossings[j].idx, db, db2),
-                              osh_kernel_step_energy(st->de, crossings[j].path_len, score_len));
-        }
-    }
-    return OSH_OK;
-}
-
-/**
- * @brief Accumulate fluence [1/cm²] into the FLUENCE scorer pages.
- *
- * Deposits the track length per crossing; the estimator postprocess divides by
- * the per-bin volume (geo->bin_vol_inv) once, yielding fluence [1/cm²].
- */
-enum osh_status score_step_fluence(struct osh_scoring_runtime const *rt,
-                                   struct osh_scoring_accumulator *acc_set,
-                                   struct osh_scoring_geometry_score_group const *group,
-                                   struct osh_voxel_crossing const *crossings,
-                                   size_t ncross,
-                                   struct particle const *part,
-                                   struct step const *st,
-                                   double score_len) {
-    size_t i;
-    size_t j;
-    size_t db;
-    size_t db2;
-    struct osh_scoring_page_runtime const *page;
-    struct osh_scoring_accumulator *acc;
-    (void) score_len; /* fluence books raw track length; ÷volume is done in postprocess */
-
-    for (i = 0; i < group->npages; ++i) {
-        page = &rt->pages[group->first_page + i];
-        acc = &acc_set[group->first_page + i];
-        if (!osh_scoring_page_passes_filters(page, part, st)) {
-            continue;
-        }
-        if (!resolve_diff_bins(page, rt, part, st, &db, &db2)) {
-            continue;
-        }
-        for (j = 0; j < ncross; ++j) {
-            if (crossings[j].idx >= page->diff_stride) {
-                return OSH_ESTATE;
-            }
-            osh_score_deposit(
-                acc->data, flat_bin(page, crossings[j].idx, db, db2), osh_kernel_step_fluence(crossings[j].path_len));
-        }
-    }
-    return OSH_OK;
-}
-
-/**
- * @brief Accumulate dose [MeV/g] into the DOSE scorer pages.
- *
- * @c vol_inv is read per-crossing from @c crossings[j].vol_inv (pre-filled by
- * the caller: uniform scalar for Mesh; per-R-bin LUT for Cyl).
- *
- * When @c mat_tables is available and a Settings block specifies a medium,
- * applies a stopping-power ratio correction S(ovr,E)/S(tr,E) for
- * dose-to-medium scoring.  Pure density overrides do not change the dose
- * (Fano theorem).
- */
-enum osh_status score_step_dose(struct osh_scoring_runtime const *rt,
-                                struct osh_scoring_accumulator *acc_set,
-                                struct osh_scoring_geometry_score_group const *group,
-                                struct osh_voxel_crossing const *crossings,
-                                size_t ncross,
-                                struct particle const *part,
-                                struct step const *st,
-                                double score_len) {
-    size_t i;
-    size_t j;
-    size_t db;
-    size_t db2;
-    double base_scale;
-    double dose_scale;
-    struct dose_sp_ctx sp;
-    struct osh_scoring_page_runtime const *page;
-    struct osh_scoring_accumulator *acc;
-
-    if (!(st->rho > 0.0)) {
-        return OSH_OK;
-    }
-    /* base_scale = de / (rho * score_len).  ÷volume is deferred to postprocess
-     * (accumulates in [MeV·cm³/g]; postprocess yields [MeV/g], DOSEGY then × Gy). */
-    base_scale = st->de / (score_len * st->rho);
-    sp = dose_sp_gather(rt, part, st);
-
-    /* group is a run of DOSE or DOSEGY pages for this geometry, not a run of
-     * bins.  Each page may have different filters, Settings overrides, and
-     * differential axes, so dose_scale and diff bins are page-local. */
-    for (i = 0; i < group->npages; ++i) {
-        page = &rt->pages[group->first_page + i];
-        acc = &acc_set[group->first_page + i];
-        if (!osh_scoring_page_passes_filters(page, part, st)) {
-            continue;
-        }
-        /* Dose-to-medium: scale by S(ovr)/S(tr) (1.0 when no override). */
-        dose_scale = base_scale * dose_sp_ratio(&sp, rt, page);
-        if (!resolve_diff_bins(page, rt, part, st, &db, &db2)) {
-            continue;
-        }
-        for (j = 0; j < ncross; ++j) {
-            if (crossings[j].idx >= page->diff_stride) {
-                return OSH_ESTATE;
-            }
-            osh_score_deposit(acc->data,
-                              flat_bin(page, crossings[j].idx, db, db2),
-                              osh_kernel_step_dose(crossings[j].path_len, dose_scale));
-        }
-    }
-    return OSH_OK;
-}
-
-/**
- * @brief Point-deposit counterpart of score_step_energy.
- *
- * Books the whole released energy (st->de) at the single located bin — a point has
- * no track, so there is no path fraction.  Shares the page/filter/diff machinery.
- */
-enum osh_status score_point_energy(OSH_SCORING_DEPOSIT_PARAMS) {
-    size_t i;
-    size_t j;
-    size_t db;
-    size_t db2;
-    struct osh_scoring_page_runtime const *page;
-    struct osh_scoring_accumulator *acc;
-    (void) score_len; /* a point deposit has no track length */
-
-    for (i = 0; i < group->npages; ++i) {
-        page = &rt->pages[group->first_page + i];
-        acc = &acc_set[group->first_page + i];
-        if (!osh_scoring_page_passes_filters(page, part, st)) {
-            continue;
-        }
-        if (!resolve_diff_bins(page, rt, part, st, &db, &db2)) {
-            continue;
-        }
-        for (j = 0; j < ncross; ++j) {
-            if (crossings[j].idx >= page->diff_stride) {
-                return OSH_ESTATE;
-            }
-            osh_score_deposit(acc->data, flat_bin(page, crossings[j].idx, db, db2), osh_kernel_point_energy(st->de));
-        }
-    }
-    return OSH_OK;
-}
-
-/**
- * @brief Point-deposit counterpart of score_step_dose.
- *
- * Books de/ρ (× dose-to-medium SP ratio) at the single located bin — no path.
- * Neutral particles (e.g. a de-excitation gamma) release energy here but carry it
- * away without depositing dose locally, so they book energy but never dose.
- */
-enum osh_status score_point_dose(OSH_SCORING_DEPOSIT_PARAMS) {
-    size_t i;
-    size_t j;
-    size_t db;
-    size_t db2;
-    double sp_ratio;
-    struct dose_sp_ctx sp;
-    struct osh_scoring_page_runtime const *page;
-    struct osh_scoring_accumulator *acc;
-    (void) score_len; /* a point deposit has no track length */
-
-    if (part->charge == 0) {
-        return OSH_OK;
-    }
-    if (!(st->rho > 0.0)) {
-        return OSH_OK;
-    }
-    sp = dose_sp_gather(rt, part, st);
-
-    /* group is a run of DOSE or DOSEGY pages for the same geometry, not a run of
-     * bins.  Each page may have different filters, Settings overrides, and
-     * differential axes, so the page loop remains even though a point deposit
-     * normally has only one located crossing. */
-    for (i = 0; i < group->npages; ++i) {
-        page = &rt->pages[group->first_page + i];
-        acc = &acc_set[group->first_page + i];
-        if (!osh_scoring_page_passes_filters(page, part, st)) {
-            continue;
-        }
-        sp_ratio = dose_sp_ratio(&sp, rt, page);
-        if (!resolve_diff_bins(page, rt, part, st, &db, &db2)) {
-            continue;
-        }
-        for (j = 0; j < ncross; ++j) {
-            if (crossings[j].idx >= page->diff_stride) {
-                return OSH_ESTATE;
-            }
-            osh_score_deposit(
-                acc->data, flat_bin(page, crossings[j].idx, db, db2), osh_kernel_point_dose(st->de, st->rho, sp_ratio));
-        }
-    }
-    return OSH_OK;
-}
-
-/**
- * @brief Accumulate dose-averaged LET [MeV/cm] via a two-pass accumulator.
- *
- * Uses S(medium,E)·ρ from the SP tables when available; falls back to de/score_len.
- * Per-page Settings overrides apply S(ovr,E)·ρ_ovr (medium) or S(tr,E)·ρ_ovr (density-only).
- */
-enum osh_status score_step_dlet(struct osh_scoring_runtime const *rt,
-                                struct osh_scoring_accumulator *acc_set,
-                                struct osh_scoring_geometry_score_group const *group,
-                                struct osh_voxel_crossing const *crossings,
-                                size_t ncross,
-                                struct particle const *part,
-                                struct step const *st,
-                                double score_len) {
-    size_t i;
-    size_t j;
-    double sp_transport;
-    double let_default;
-    double let_step;
-    double mean_energy;
-    double e_per_nuc;
-    double rho_ovr; /* density used for the per-page LET override */
-    size_t proj_idx;
-    int have_proj;
-    struct osh_scoring_page_runtime const *page;
-    struct osh_scoring_accumulator *acc;
-    struct osh_scoring_page_override const *sset; /* per-page settings override pointer */
-    struct osh_material_runtime const *mat_tables = rt->mat_tables;
-    e_per_nuc = 0.0; /* initialized to satisfy MSVC C4701; overwritten when table-based projectile data is available */
-
-    /* LET is only defined for charged particles in a material (not vacuum). */
-    if (part->z == 0 || part->a == 0 || !(st->rho > 0.0)) {
-        return OSH_OK;
-    }
-
-    /* Compute table-based LET in the transport medium: S(medium,E)·ρ [MeV/cm].
-     * sp_transport is kept separately so density-only overrides can reuse it.
-     * Fall back to geometric de/score_len when tables are unavailable. */
-    have_proj = 0;
-    proj_idx = 0;
-    sp_transport = 0.0;
-    let_default = st->de / score_len; /* geometric fallback */
-    if (mat_tables && st->medium >= 0) {
-        mean_energy = 0.5 * (st->p[3] + st->q[3]);
-        e_per_nuc = mean_energy / (double) part->a;
-        if (find_proj_idx(mat_tables, (unsigned int) part->z, &proj_idx)) {
-            have_proj = 1;
-            sp_transport = osh_material_runtime_sp_lookup(mat_tables, (size_t) st->medium, proj_idx, e_per_nuc);
-            let_default = sp_transport * st->rho;
-        }
-    }
-
-    for (i = 0; i < group->npages; ++i) {
-        page = &rt->pages[group->first_page + i];
-        acc = &acc_set[group->first_page + i];
-        if (!osh_scoring_page_passes_filters(page, part, st)) {
-            continue;
-        }
-        let_step = let_default;
-        if (have_proj && page->has_sset) {
-            sset = &page->sset;
-            if (sset->has_medium && sset->medium >= 0) {
-                rho_ovr = sset->has_density_g_cm3 ? sset->density_g_cm3 : mat_tables->rho[sset->medium];
-                let_step =
-                    osh_material_runtime_sp_lookup(mat_tables, (size_t) sset->medium, proj_idx, e_per_nuc) * rho_ovr;
-            } else if (sset->has_density_g_cm3) {
-                let_step = sp_transport * sset->density_g_cm3;
-            }
-        }
-        for (j = 0; j < ncross; ++j) {
-            if (crossings[j].idx >= page->diff_stride) {
-                return OSH_ESTATE;
-            }
-            /* Two-pass dose-averaged LET:
-             *   data  = sum(LET · dose_weight)   [MeV/cm · MeV]
-             *   data2 = sum(dose_weight)          [MeV]
-             * osh_scoring_postprocess() divides data by data2 to yield LETd. */
-            deposit_two_pass(
-                acc, crossings[j].idx, let_step, osh_kernel_dose_weight(st->de, crossings[j].path_len, score_len));
-        }
-    }
-    return OSH_OK;
-}
-
-/**
- * @brief Accumulate track-averaged LET [MeV/cm] via a two-pass accumulator.
- *
- * Same table lookup as score_step_dlet(); uses track-length ds_vox as the weight
- * rather than dose weight.
- */
-enum osh_status score_step_tlet(struct osh_scoring_runtime const *rt,
-                                struct osh_scoring_accumulator *acc_set,
-                                struct osh_scoring_geometry_score_group const *group,
-                                struct osh_voxel_crossing const *crossings,
-                                size_t ncross,
-                                struct particle const *part,
-                                struct step const *st,
-                                double score_len) {
-    size_t i;
-    size_t j;
-    double sp_transport;
-    double let_default;
-    double let_step;
-    double mean_energy;
-    double e_per_nuc;
-    double rho_ovr; /* density used for the per-page LET override */
-    size_t proj_idx;
-    int have_proj;
-    struct osh_scoring_page_runtime const *page;
-    struct osh_scoring_accumulator *acc;
-    struct osh_scoring_page_override const *sset; /* per-page settings override pointer */
-    struct osh_material_runtime const *mat_tables = rt->mat_tables;
-    e_per_nuc = 0.0; /* initialized to satisfy MSVC C4701; overwritten when table-based projectile data is available */
-
-    /* LET is only defined for charged particles in a material (not vacuum). */
-    if (part->z == 0 || part->a == 0 || !(st->rho > 0.0)) {
-        return OSH_OK;
-    }
-
-    /* Table-based LET in transport medium; geometric fallback when unavailable. */
-    have_proj = 0;
-    proj_idx = 0;
-    sp_transport = 0.0;
-    let_default = st->de / score_len;
-    if (mat_tables && st->medium >= 0) {
-        mean_energy = 0.5 * (st->p[3] + st->q[3]);
-        e_per_nuc = mean_energy / (double) part->a;
-        if (find_proj_idx(mat_tables, (unsigned int) part->z, &proj_idx)) {
-            have_proj = 1;
-            sp_transport = osh_material_runtime_sp_lookup(mat_tables, (size_t) st->medium, proj_idx, e_per_nuc);
-            let_default = sp_transport * st->rho;
-        }
-    }
-
-    for (i = 0; i < group->npages; ++i) {
-        page = &rt->pages[group->first_page + i];
-        acc = &acc_set[group->first_page + i];
-        if (!osh_scoring_page_passes_filters(page, part, st)) {
-            continue;
-        }
-        let_step = let_default;
-        if (have_proj && page->has_sset) {
-            sset = &page->sset;
-            if (sset->has_medium && sset->medium >= 0) {
-                rho_ovr = sset->has_density_g_cm3 ? sset->density_g_cm3 : mat_tables->rho[sset->medium];
-                let_step =
-                    osh_material_runtime_sp_lookup(mat_tables, (size_t) sset->medium, proj_idx, e_per_nuc) * rho_ovr;
-            } else if (sset->has_density_g_cm3) {
-                let_step = sp_transport * sset->density_g_cm3;
-            }
-        }
-        for (j = 0; j < ncross; ++j) {
-            if (crossings[j].idx >= page->diff_stride) {
-                return OSH_ESTATE;
-            }
-            /* Two-pass track-averaged LET:
-             *   data  = sum(LET · ds_vox)   [MeV/cm · cm = MeV]
-             *   data2 = sum(ds_vox)         [cm]
-             * osh_scoring_postprocess() divides data by data2 to yield LETt. */
-            deposit_two_pass(
-                acc, crossings[j].idx, let_step, osh_kernel_track_weight(st->ds, crossings[j].path_len, score_len));
-        }
-    }
-    return OSH_OK;
-}
-
-/**
- * @brief Accumulate dose-averaged (z_eff/β)² [dimensionless] via a two-pass accumulator.
- *
- * Requires mat_tables for the rest mass needed to compute β.  Neutrals and
- * particles without SP table entries contribute nothing.
- *
- * @ref Kalholm et al. Medical Physics. 2023 Jan;50(1):651-9.
- *      https://doi.org/10.1002/mp.16029
- *
- */
-enum osh_status score_step_dqeff(struct osh_scoring_runtime const *rt,
-                                 struct osh_scoring_accumulator *acc_set,
-                                 struct osh_scoring_geometry_score_group const *group,
-                                 struct osh_voxel_crossing const *crossings,
-                                 size_t ncross,
-                                 struct particle const *part,
-                                 struct step const *st,
-                                 double score_len) {
-    size_t i;
-    size_t j;
-    double mean_energy;
-    double beta;
-    double qeff;
-    size_t proj_idx;
-    struct osh_scoring_page_runtime const *page;
-    struct osh_scoring_accumulator *acc;
-    struct osh_material_runtime const *mat_tables = rt->mat_tables;
-
-    if (part->z == 0) {
-        return OSH_OK;
-    }
-    if (!mat_tables) {
-        return OSH_OK;
-    }
-    if (!find_proj_idx(mat_tables, (unsigned int) part->z, &proj_idx)) {
-        return OSH_OK;
-    }
-
-    mean_energy = 0.5 * (st->p[3] + st->q[3]);
-    beta = particle_beta(mean_energy, mat_tables->projectile_mass_mev[proj_idx]);
-    if (!(beta > 0.0)) {
-        return OSH_OK;
-    }
-    qeff = particle_qeff((int) part->z, beta);
-
-    for (i = 0; i < group->npages; ++i) {
-        page = &rt->pages[group->first_page + i];
-        acc = &acc_set[group->first_page + i];
-        if (!osh_scoring_page_passes_filters(page, part, st)) {
-            continue;
-        }
-        for (j = 0; j < ncross; ++j) {
-            if (crossings[j].idx >= page->diff_stride) {
-                return OSH_ESTATE;
-            }
-            /* Two-pass dose-averaged (z_eff/β)²:
-             *   data  = sum(qeff · dose_weight)
-             *   data2 = sum(dose_weight)           [MeV]
-             * osh_scoring_postprocess() divides to yield the dose-averaged value. */
-            deposit_two_pass(
-                acc, crossings[j].idx, qeff, osh_kernel_dose_weight(st->de, crossings[j].path_len, score_len));
-        }
-    }
-    return OSH_OK;
-}
-
-/**
- * @brief Accumulate track-averaged (z_eff/β)² [dimensionless] via a two-pass accumulator.
- *
- * Same as score_step_dqeff() but uses track-length ds_vox as the weight.
- *
- * @ref Kalholm et al. Medical Physics. 2023 Jan;50(1):651-9.
- *      https://doi.org/10.1002/mp.16029
- *
- */
-enum osh_status score_step_tqeff(struct osh_scoring_runtime const *rt,
-                                 struct osh_scoring_accumulator *acc_set,
-                                 struct osh_scoring_geometry_score_group const *group,
-                                 struct osh_voxel_crossing const *crossings,
-                                 size_t ncross,
-                                 struct particle const *part,
-                                 struct step const *st,
-                                 double score_len) {
-    size_t i;
-    size_t j;
-    double mean_energy;
-    double beta;
-    double qeff;
-    size_t proj_idx;
-    struct osh_scoring_page_runtime const *page;
-    struct osh_scoring_accumulator *acc;
-    struct osh_material_runtime const *mat_tables = rt->mat_tables;
-
-    if (part->z == 0) {
-        return OSH_OK;
-    }
-    if (!mat_tables) {
-        return OSH_OK;
-    }
-    if (!find_proj_idx(mat_tables, (unsigned int) part->z, &proj_idx)) {
-        return OSH_OK;
-    }
-
-    mean_energy = 0.5 * (st->p[3] + st->q[3]);
-    beta = particle_beta(mean_energy, mat_tables->projectile_mass_mev[proj_idx]);
-    if (!(beta > 0.0)) {
-        return OSH_OK;
-    }
-    qeff = particle_qeff((int) part->z, beta);
-
-    for (i = 0; i < group->npages; ++i) {
-        page = &rt->pages[group->first_page + i];
-        acc = &acc_set[group->first_page + i];
-        if (!osh_scoring_page_passes_filters(page, part, st)) {
-            continue;
-        }
-        for (j = 0; j < ncross; ++j) {
-            if (crossings[j].idx >= page->diff_stride) {
-                return OSH_ESTATE;
-            }
-            /* Two-pass track-averaged (z_eff/β)²:
-             *   data  = sum(qeff · ds_vox)
-             *   data2 = sum(ds_vox)           [cm]
-             * osh_scoring_postprocess() divides to yield the track-averaged value. */
-            deposit_two_pass(
-                acc, crossings[j].idx, qeff, osh_kernel_track_weight(st->ds, crossings[j].path_len, score_len));
-        }
-    }
     return OSH_OK;
 }
